Style product item component
