--- src/containers/CartItem.css
@@ -1,47 +1,0 @@
-.cart-item {
-    position: relative;
-    display: flex;
-    flex-direction: row;
-    padding: 2rem;
-    margin: auto;
-    transition: all .2s;
-}
-
-.cart-item-col {
-    display: flex;
-    flex-direction: column;
-    justify-content: space-evenly;
-    margin-right: 1rem;
-}
-
-.cart-item-img {
-    height: 8rem;
-    width: auto;
-    margin-right: 1rem;
-}
-
-@media only screen and (max-width: 975px) {
-    .cart-item {
-        min-width: 50vw;
-    }
-}
-@media only screen and (max-width: 800px) {
-    .cart-item {
-        min-width: 60vw;
-    }
-}
-@media only screen and (max-width: 650px) {
-    .cart-item {
-        min-width: 75vw;
-    }
-}
-
-@media only screen and (max-width: 475px) {
-    .cart-item {
-        align-items: center;
-        width: 100vw;
-    }
-    .cart-item-img {
-        height: 9rem;
-    }
-}
